fwbot/pyproject.toml
[tool.poetry]
name = "fwbot"
version = "0.1.0"
description = ""
authors = ["Anastasiia Maiorova <anastasia.mayorova.work@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
langchain = "^0.1.17"
langchain-openai = "^0.1.4"
langchainhub = "^0.1.15"
python-telegram-bot = "^21.1.1"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.127.0"
google-cloud-speech = "^2.26.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "fwbot.start_up:start_up"
